+++ geopackage-sdk/src/main/java/mil/nga/geopackage/extension/style/Styles.java
@@ -27,6 +27,28 @@ public class Styles {
     private Map<GeometryType, StyleRow> styles = new HashMap<>();
 
     /**
+     * Table styles flag
+     */
+    private boolean tableStyles;
+
+    /**
+     * Constructor
+     */
+    public Styles() {
+        this(false);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param tableStyles table styles
+     * @since 3.5.0
+     */
+    public Styles(boolean tableStyles) {
+        this.tableStyles = tableStyles;
+    }
+
+    /**
      * Set the default style icon
      *
      * @param styleRow default style
@@ -42,6 +64,9 @@ public class Styles {
      * @param geometryType geometry type
      */
     public void setStyle(StyleRow styleRow, GeometryType geometryType) {
+        if (styleRow != null) {
+            styleRow.setTableStyle(tableStyles);
+        }
         if (geometryType != null) {
             if (styleRow != null) {
                 styles.put(geometryType, styleRow);
@@ -129,6 +154,26 @@ public class Styles {
      */
     public boolean hasDefault() {
         return defaultStyle != null;
+    }
+
+    /**
+     * Is table styles
+     *
+     * @return table styles flag
+     * @since 3.5.0
+     */
+    public boolean isTableStyles() {
+        return tableStyles;
+    }
+
+    /**
+     * Set table styles flag
+     *
+     * @param tableStyles table styles flag
+     * @since 3.5.0
+     */
+    public void setTableStyles(boolean tableStyles) {
+        this.tableStyles = tableStyles;
     }
 
 }