+++ geopackage-sdk/src/main/java/mil/nga/geopackage/features/user/ManualFeatureQuery.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.Map;
 
 import mil.nga.geopackage.BoundingBox;
+import mil.nga.geopackage.db.CoreSQLUtils;
 import mil.nga.geopackage.features.index.FeatureIndexManager;
 import mil.nga.sf.GeometryEnvelope;
 import mil.nga.sf.proj.Projection;
@@ -124,7 +125,10 @@ public class ManualFeatureQuery {
      * @return count
      */
     public int countWithGeometries() {
-        return featureDao.count(featureDao.getGeometryColumnName() + " IS NOT NULL", null);
+        return featureDao.count(
+                CoreSQLUtils.quoteWrap(featureDao.getGeometryColumnName())
+                        + " IS NOT NULL",
+                null);
     }
 
     /**