+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * XYZ tile rebranding, previously referred to as Google
 * Separation of columns from user tables
 * Queries by specified columns
+* Feature Preview for drawing a single preview tile of features
+* Feature Style Extension fix to use an individual feature style over a default table icon
+* GeoPackageImpl and FeatureIndexGeoPackageResults classes changed to be public
 * gradle plugin updated to 3.5.3
 
 ## [3.4.0](https://github.com/ngageoint/geopackage-android/releases/tag/3.4.0) (11-14-2019)